Revert "Merge branch 'main' of https://github.com/azerty0004/PiDev"
This reverts commit 914a060bfd3a22cba274aaac1cdab95bae1c0599, reversing
changes made to f57a6a91696688c1f158ee5de1aaa3670ba80e99.

+++ src/main/java/tn/esprit/infini/Pidev/Services/ITransaction.java
@@ -12,8 +12,8 @@ public interface ITransaction {
 
   Transaction updateTransaction(Transaction transaction);
 
-   Transaction retrieveTransaction(Long idTransaction);
+   Transaction retrieveTransaction(Integer idTransaction);
 
-    void deleteTransaction(Long idTransaction);
+    void deleteTransaction(Integer idTransaction);
 
 }
